pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
python-telegram-bot = "^22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
